Parser/pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = ""
authors = ["Andre Borges Medeiros <andborges@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
aio-pika = "^9.4.2"
protobuf = "^5.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
